requirements.txt
python-telegram-bot==22.3
pytz
reportlab
python-dotenv
flask
nest-asyncio
psycopg2-binary
mysql-connector-python
gunicorn 